+++ CHANGELOG.md
@@ -11,6 +11,8 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 ### New
 * Update installation and contributing docs
 * Meltano implement role-based access control - [!368](https://gitlab.com/meltano/meltano/merge_requests/368)
+* Add deep linking to dashboards
+* Add deep linking to reports
 
 ### Changes
 
@@ -19,7 +21,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 * Fixed dashboards initial load issue due to legacy (and empty) `carbon.dashboard.m5o` file
 
 ### Breaks
-
+* New standardized approach for `.m5o` id generation (will need to remove any dashboard.m5o and report.m5o)
 
 ## 0.11.0 - (2019-02-19)
 ---